🏗 Change the 'ChangePW' component as a screen.
