+++ src/components/screen/Setting/styles.ts
@@ -1,4 +1,3 @@
-import TextInput from '../../shared/TextInput';
 import styled from 'styled-components/native';
 
 export const Container = styled.View`
@@ -44,28 +43,3 @@ export const ItemLabel = styled.Text`
   font-size: 16px;
   flex: 1;
 `;
-
-export const ModalHeader = styled.View`
-  height: 52px;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const ModalTitle = styled.Text`
-  color: ${({ theme }): string => theme.fontColor};
-  font-size: 20px;
-`;
-
-export const ModalCloseButton = styled.TouchableOpacity`
-  position: absolute;
-  right: 10px;
-`;
-
-export const InnerContainer = styled.View`
-  padding: 15px;
-  flex: 1;
-`;
-
-export const StyledTextInput = styled(TextInput).attrs(({ theme }) => ({
-  placeholderTextColor: theme.placeholder,
-}))``;
